+++ NEWS.md
@@ -1,3 +1,7 @@
+# pmparser 1.0.8
+* Increased timeout for downloading citation file.
+* Reordered steps to check citation table even if other tables are up-to-date.
+
 # pmparser 1.0.7
 * Explicitly added `curl` to Imports, since it's used by `utils::download.file()`.
 * Set "HUGE" option for `xml2::read_xml()` to avoid occassional error.